[BUG] Avoid code execution race condition with common gpt.py
Previously, an individual's code was written to the gpt.py file adjacent
to the problem's eval.py script. Then, a subprocess was created to run
the eval.py script. This led to issues with the wrong code getting
executed due to race conditions. Now, we write the code to the
individual's directory and provide the code path to the eval.py script.
Furthermore, we clean up the eval.py scripts:

- Organize imports
- Use llamda.utils instead of sys hack to access dynamic loader utils
- Use argparse